calvin ok

+++ common/MyMove.h
@@ -1213,7 +1213,7 @@ namespace star
                 total_move_plans.push_no_wait(move_step);
             }
             
-            for(int i = 0 ; i < context.coordinator_num; i ++ ){
+            for(size_t i = 0 ; i < context.coordinator_num; i ++ ){
                 LOG(INFO) << i << " : " << node_load[i];
             }
